Seems to work as intended.

+++ backend/server/models/week.js
@@ -2,7 +2,7 @@ module.exports = (sequelize, DataTypes) => {
   const Week = sequelize.define(
     'Week',
     {
-      points: DataTypes.INTEGER,
+      points: DataTypes.DOUBLE,
       weekNumber: DataTypes.INTEGER,
       feedback: DataTypes.STRING
     },
